Found an error in QOS5 handling that required a refactor: Factored out ST_Layer_Shared with X_Eliminate functions that are used by both layers. This makes the Timelayer independent of the Secondlayer (it now only uses ST_Layer_Shared).
For QOS6 and QOS7 some additional issues were found: QOS6 can currently output times which run counter to the received data causing subsequent QOS7/8 handling to to become active despite this probably being needed. Additionally, QOS7 is implemented wrongly because the code assumes that Prev points to the previous minute whereas it correctly points to a previous telegram from previous minute tens. QOS6 either needs to be adjusted to not output anything counter to the received telegrams (which may make it impossible to test in the current setup!) or alternatively be dropped altogether because the cases where QOS6 works but QOS5 does not seem to be negligible to the point that it is not possible to construct an adequate test case?

Additionally, for QOS7 some computation to create a virtual “prev minute” telegram is required. It can be based off the code from QOS6 except we'd add +9 rather than +10minutes or something? Additionally, the code currently writes directly to the Prev telegram which may not be desirable if it turns out that this case is not to be used.

S = Debugging needs to continue for QOS6 and QOS7 and then some other tests wrt. DST transitions could be added.