Bitlayer: Fix bug where clock would get out of sync after ~2 weeks of uptime
A bug was observed where the clock would get out of sync after a certain
(rather long) uptime.

The bug would manifest as follows:

At end of minute, in place of 300 or 301, signals 003 or 031 would be
recognized, with the `3` being reported one or two seconds too late depending
on the next byte (for 0 would commonly appear two seconds late, for 1 would
appear one second late).

It is suspected that this was due to a drift in the time: The previous code
would always increment the time by a constant number in order to compute the
expiry of the seconds. As the clock uptime progressed, this could get out of
sync (DCF time reported /= computed second increments). As a result, the timer
for detecting a “no signal” condition at the end of minute would not expire in
time, rather the ”0” from the beginning of the next minute would be detected
and since that would shift the timer again += 1sec, a follow-up ”0” might still
be recognized before finally giving space to the “3”.

Effectively the continuous maintenance of a timer by incrementing it with a
fixed amount was the error here.

This commit attempts to fix this as follows:

Instead of incrementing by a constant amount, whenever a signal is recognized,
the timer is set to expire at the beginning of that signal + 1sec. The actual
triggering of the timer is delayed by 300ms which gives enough space to detect
the “end” of the next seconds' signal.

Note that this timeout must not be set too high as not to arrive at the same
error again, but it seems that the limit for that is anything (well) below 2 sec
in total i.e. 1.3sec which is newly the “maximum wait time” should be safe in
many circumstances.

There is still the issue that if there is some sort of fluctuation in the delays
one may quickly exceed this limit and something like `31` or `30` is then
generated in place of a proper `1` or `0`.

Running the clock for prolonged time might help with finding out if this should
still be a concern. If it turns out to be an issue it may also make sense to
rewrite ISR + Bitlayer to follow a different model where the bitlayer has some
awareness about the “ongoing” signal and might thus monitor the timeout and
synchronization even more closely.