Intermeidate commit: Rework the Bitlayer
The previous Bitlayer had some serious issues mostly resulting from a bad
separation of concerns between ISR and Bitlayer.

This commit introduces a change in Bitlayer and Low Level to make the ISR
more robust and implement advanced processing on the Bitlayer that is supposed
to lock to the phase more reliably.

This is implemented using three state machines, the first two of which are
responsible for validating the incoming bits as reported by the ISR. The third
“main” state machine is then responsible for identifying the bits from the
preprocessed ISR values.

The data format for the low-level trasmission of GPIO query results between the
Low Level and the Bitlayer is newly an unsigned integer wtih a leading `1` to
identify “start of contents” followed by the individual GPIO querying results
at constant intervals of 7ms as read from inside the ISR.

The first state machine then goes on to flatten the signal and eliminates any
single 0/1 values that may have been read wrongly. The second state machine then
goes about to detect whether there is a pattern like “1 only” (INPUT_1) or
“1 followed by zeroes” (X0) or ...

The third state machine then maintains how far the processing has progressed
inside the current second (it is called ten times per second). Instead of
relying on the internal clock for measurement at this stage, the new code
assumes that the ticker (integrated with the Bitlayer) does a sufficiently good
job and moves between the states at constant tick intervals, emitting on average
one “reading” once every 10 invocations (equiv every 1 sec).

In order to document the state machines, graphs have been drawn into the
code comments and for the main state machine a dedicated “blstatechart.pdf”
document is provided.

This code is currently mostly untested. Upcoming changes are supposed to improve
introspection through the GUI which may help to evaluate whether the code
behaves according to specification in the good cases. Further tests are then
intended to be performed using real-world use case.